Only require libphonenumber when rendering phone inputs
